# controllers/user_controller.go
package controllers

/* Generated by egg v0.0.1 */

import (
	handlers "__EGG_NAMESPACE__/models/handlers/user_handlers"
	"__EGG_NAMESPACE__/services"
	"github.com/labstack/echo/v4"
)

type UserController struct {
	Name      string
	Registrar *services.Registrar
}

func (uc UserController) ControllerName() string {
	return uc.Name
}

func BuildUserController(p *services.Registrar) UserController {
	return UserController{
		Name:      "/users",
		Registrar: p,
	}
}

// @Summary Delete User by their UUID
// @Description get string by ID
//
// @ID          DeleteUserByUUID
// @Tags        Users
// @Produce     json
// @Param       user_id             path         string                         true "User Id"          default("e38e78a4-2ca3-4c59-a3ea-a2019866e593")
// @Param       Authorization       header       string                         true "admin header"     default("Bearer token")
// @Success     200                 {object}     responses.DeleteUserResponse
// @Router      /users/{user_id}    [delete]
func (UserController *UserController) DeleteUser(ctx echo.Context) error {
	return handlers.DeleteUserJsonHandler(ctx, UserController.Registrar)
}

// @Summary Signup to the app
// @Description Signup using the requests.NewUserRequest
//
// @ID          Signup
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       SignupRequest   body        NewUserRequest          true "Signup Request"
// @Success     200             {object}    responses.LoginResponse
// @Failure     400             {object}    responses.LoginResponse
// @Failure     500             {object}    responses.LoginResponse
// @Router      /users/signup   [post]
func (UserController *UserController) Signup(ctx echo.Context) error {
	return handlers.RegisterJsonHandler(ctx, UserController.Registrar)
}

// @Summary Login
// @Description to a user account with either email or username
//
// @ID          Login
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       LoginRequest body           LoginRequest            true "Log in request"
// @Success     200             {object}    responses.LoginResponse
// @Failure     400             {object}    responses.LoginResponse
// @Failure     401             {object}    responses.LoginResponse
// @Failure     500             {object}    responses.LoginResponse
// @Router      /users/login    [post]
func (uc *UserController) Login(ctx echo.Context) error {
	return handlers.LoginJsonHandler(ctx, uc.Registrar)
}

// @Summary Get Current User
// @Description Get the Current User by the uuid storred in the Claims header
//
// @ID          GetCurrentLoggedInUser
// @Tags        Users
// @Produce     json
// @Param       authorization   header       string                         true "admin header"     default(Bearer token)
// @Success     200             {object}     responses.UserResponse
// @Failure     400             {object}     responses.UserResponse
// @Failure     401             {object}     responses.UserResponse
// @Router      /users/current  [get]
func (UserController *UserController) GetCurrent(ctx echo.Context) error {
	return handlers.CurrentJsonHandler(ctx, UserController.Registrar)
}

// @Summary		Upload file
// @Description	Upload file
//
// @ID				UploadProfilePicture
// @Tags            Users
// @Accept			multipart/form-data
// @Produce			json
// @Param			file				formData	file			true	"this is a test file"
// @Param           authorization		header      string          true	"admin header"        default(Bearer token)
// @Success		    200					{string}	UserResponse
// @Failure		    400					{object}	UserResponse
// @Failure		    404					{object}	UserResponse
// @Failure		    404					{object}	UserResponse
// @Router			/users/profile		[post]
func (UserController *UserController) UploadProfilePicture(ctx echo.Context) error {
	return handlers.UploadProfilePictureJsonHandler(ctx, UserController.Registrar)
}

// @Summary Get User Profile by Authorization Header
// @Description Get User Profile by Authorization Header
//
// @ID          GetProfilePicture
// @Tags        Users
// @Produce     json
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Success     200                 {object}     responses.StringResponse
// @Failure     401                 {object}     responses.StringResponse
// @Failure     403                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /users/profile		[get]
func (uc *UserController) GetProfile(ctx echo.Context) error {
	return handlers.GetProfilePictureJsonHandler(ctx, uc.Registrar)
}

// @Summary Get All Users
// @Description Get All Users. Must be Admin using the new mediator pattern
//
// @ID          GetUsers
// @Tags        Users
// @Produce     json
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Success     200                 {object}     UsersResponse
// @Failure     403                 {object}     UsersResponse
// @Failure     404                 {object}     UsersResponse
// @Failure     500                 {object}     UsersResponse
// @Router      /users/			    [get]
func (uc *UserController) GetUsers(ctx echo.Context) error {
	return handlers.GetUsersJsonHandler(ctx, uc.Registrar)
}

// @Summary Update User Credentials
// @Description Update User Credentials by id, Must be the same user in the jwt
// @Description that is being updated
//
// @ID          UpdateUser
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       UpdateCredentialsRequest body         UpdateCredentialsRequest       true "Update User Credentials Request"
// @Param       Authorization            header       string                         true "admin header"                    default(Bearer token)
// @Success     200                      {object}     UpdateUserResponse
// @Failure     400                      {object}     UpdateUserResponse
// @Failure     401                      {object}     UpdateUserResponse
// @Failure     500                      {object}     UpdateUserResponse
// @Router      /users/creds             [post]
func (uc *UserController) UpdateUser(ctx echo.Context) error {
	return handlers.UpdateUserJsonHandler(ctx, uc.Registrar)
}

func (uc UserController) Attatch(e *echo.Echo, middlewares ...echo.MiddlewareFunc) {
	// Register the namespaces for the endopoints
	api := e.Group("/api" + uc.Name)
	api.GET("", uc.GetUsers, middlewares...)
	api.POST("/login", uc.Login)
	api.POST("/signup", uc.Signup)
	api.GET("/current", uc.GetCurrent, middlewares...)
	api.POST("/profile", uc.UploadProfilePicture, middlewares...)
	api.POST("/creds", uc.UpdateUser, middlewares...)
	api.GET("/profile", uc.GetProfile, middlewares...)
	api.DELETE("/:user_id", uc.DeleteUser, middlewares...)
}