# models/requests/NewUserRequest.go
package requests


type NewUserRequest struct {
	Username string `json:"username"`
	Email    string `json:"email"`
	Password string `json:"password"`
	IsAdmin  bool   `json:"isAdmin"`
} // @name NewUserRequest