# models/responses/UserResponse.go
/* Generated by egg v0.0.1 */

package responses

import (
	"errors"
	"time"

	"__EGG_NAMESPACE__/db/repository"
	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
)

type UserData struct {
	ID              uuid.UUID  `json:"id"`
	Email           string     `json:"email"`
	Username        string     `json:"username"`
	CreatedDatetime *time.Time `json:"created_datetime"`
	UpdatedDatetime *time.Time `json:"updated_datetime"`
	ProfilePicUrl   *string    `json:"profile_pic_url"`
	Admin           bool       `json:"admin"`
}

type UserResponse struct {
	Data    *UserData `json:"data"`
	Success bool      `json:"success"`
	Message string    `json:"message"`
} // @name UserResponse

func UserDataFromRepository(repository *repository.User) *UserData {
	if repository == nil {
		return nil
	}
	return &UserData{
		ID:              repository.ID,
		Email:           repository.Email,
		Username:        repository.Username,
		CreatedDatetime: repository.CreatedDatetime,
		UpdatedDatetime: repository.UpdatedDatetime,
		ProfilePicUrl:   repository.ProfilePicUrl,
		Admin:           repository.Admin,
	}
}

func NewUserResponse() *UserResponse {
	return &UserResponse{Success: false, Message: ""}
}

func (UserResponse *UserResponse) Fail(ctx echo.Context, code int, err error) error {
	UserResponse.Message = err.Error()
	return ctx.JSON(code, UserResponse)
}

func (UserResponse *UserResponse) Successful(ctx echo.Context, user *repository.User) error {
	UserResponse.Data = UserDataFromRepository(user)
	UserResponse.Success = true
	return ctx.JSON(200, UserResponse)
}

func (ur *UserResponse) Component(
	ctx echo.Context,
	user *repository.User,
	code int,
	err error,
) error {
	return errors.New("oops")
}